+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
+
+## [6.13.0] - 2024-03-09
 ### Fixed
 - Fixed #177 - SFTP MoveToFile fails if destination file already exists
-
 ### Added
 - Updated all backends to use error wrapping for Read, Write, Close, Seek. #174
 - Update GS backend to comply with new io integration test suite. #172